Reduce severity of missing timestamp log messages (#1736)
If a metadata-set is indexed for a second time, it's creation time is
copied from the existing index. If no previously indexed metadata-set is
found then a message with a stack-trace is logged. Reduce the severity
of this message from warning to info and don't log the stack-trace any
more.