Stabilize tests
The problem with the indexing time is, that the call to get the
current time is done, before the index button is clicked. We have
to get the current time, after clicking on the index button in a
"then()" block. After indexing is finished, the start time of the
index process is shown, which should be equal to the current time
after clicking the index button.